+++ package.js
@@ -30,14 +30,20 @@ function configure( api ){
     api.use( 'dburles:mongo-collection-instances@0.3.6 || 1.0.0-beta300.0' );
     api.use( 'less@4.0.0', 'client' );
     api.use( 'mongo' );
+    api.use( 'pwix:accounts-hub@1.1.0' );
+    api.use( 'pwix:collection-timestampable@2.0.0' );
+    api.use( 'pwix:field@1.0.0-rc' );
     api.use( 'pwix:i18n@1.5.7' );
     api.use( 'pwix:modal-info@1.5.0' );
     api.use( 'pwix:plus-button@0.11.2 || 1.0.0-rc' );
+    api.use( 'pwix:tabular@1.0.0-rc' );
     api.use( 'pwix:toggle-switch@0.3.3' );
     api.use( 'pwix:ui-bootstrap5@2.0.0' );
+    api.use( 'pwix:ui-utils@1.3.0' );
     api.use( 'tmeasday:check-npm-versions@1.0.2 || 2.0.0-beta.0', 'server' );
     api.addFiles([
         'src/client/components/teManager/teManager.js',
+        'src/client/components/teManagerExt/teManagerExt.js',
         'src/client/components/teSerializer/teSerializer.js',
         'src/client/components/teScriber/teScriber.js',
         'src/client/components/teSwitch/teSwitch.js'