+++ src/common/collections/contents/collection.js
@@ -1,0 +1,6 @@
+/*
+ * pwix:editor/src/common/collections/contents/collection.js
+ */
+
+Editor.collections.Contents = {
+};