+++ src/common/js/index.js
@@ -4,7 +4,7 @@
 
 import './global.js';
 import './constants.js';
-//
-import './collections.js';
 import './configure.js';
 import './i18n.js';
+//
+import './collections.js';