+++ src/server/js/collections.js
@@ -2,4 +2,4 @@
  * pwix:editor/src/server/js/collections.js
  */
 
-import '../../collections/contents/server/index.js';
+import '../../common/collections/contents/server/index.js';